+++ src/main/java/com/garahe/database/UserDAO.java
@@ -5,4 +5,5 @@ import com.garahe.models.User;
 public interface UserDAO {
     boolean registerUser(User user);
     User authenticateUser(String email, String password);
+    User getUserByEmail(String email);
 }